+++ Sources/Aphrodite.cpp
@@ -48,9 +48,9 @@ vector<Insect*> Aphrodite::split(Insect* insect)
     vector<Traits*> v = inheritedTraits(insect);
     Traits* t1 = v[0];
     Traits* t2 = v[1];
-    unsigned short a1 = Organism::get_latest_organism_ID()+1;
+    unsigned a1 = Organism::get_latest_organism_ID()+1;
     Insect* d1 = new Insect(posn1,*t1,a1,insectIndex,universe,MALE);
-    unsigned short a2 = Organism::get_latest_organism_ID()+1;
+    unsigned a2 = Organism::get_latest_organism_ID()+1;
     Insect* d2 = new Insect(posn2,*t2,a2,insectIndex,universe,MALE);
     universe->killInsect(insect);
     return {d1,d2};
@@ -139,7 +139,7 @@ Insect* Aphrodite::mating(Insect* insect1,Insect* insect2)
     int del = (e1*childenergy)/(e1+e2);
     insect1->updateEnergy(e1-del);
     insect2->updateEnergy(e2-childenergy+del);
-    unsigned short a = Organism::get_latest_organism_ID()+1;
+    unsigned a = Organism::get_latest_organism_ID()+1;
     int g = rand()%2;
     Insect* offspring = new Insect(posn,v,a,insectIndex,universe,g);
     return offspring;
